Created Model

+++ model/RubickscubeBitboard.cpp
@@ -66,17 +66,17 @@ private:
         return ret;
     }
 
-    //    This function was used for testing / printing
+   // This function was used for testing / printing
 
-    //    void print5bitbin(int a){
-    //        for(int i=4; i>=0; i--){
-    //            if(a & (1 << i)) cout << 1;
-    //            else cout << 0;
-    //        }
-    //    }
+    // void print5bitbin(int a){
+    //     for(int i=4; i>=0; i--){
+    //         if(a & (1 << i)) cout << 1;
+    //         else cout << 0;
+    //     }
+    // }
 
 public:
-    uint8_t bitboard[6]{};
+    uint64_t bitboard[6]{};
 
     RubickscubeBitboard() {
         for (int side=0;side<6;side++) {
@@ -368,14 +368,14 @@ public:
 
 //        Following was used for Testing / Printing
 
-//        cout << top_front_right << " "; print5bitbin(get5bitCorner(top_front_right )); cout  << "\n";
-//        cout << top_front_left << " "; print5bitbin(get5bitCorner(top_front_left )); cout << "\n";
-//        cout << top_back_right << " "; print5bitbin(get5bitCorner(top_back_right )); cout << "\n";
-//        cout << top_back_left  << " "; print5bitbin(get5bitCorner(top_back_left  )); cout << "\n";
-//        cout << bottom_front_right  << " "; print5bitbin(get5bitCorner(bottom_front_right  )); cout << "\n";
-//        cout << bottom_front_left << " "; print5bitbin(get5bitCorner(bottom_front_left )); cout << "\n";
-//        cout << bottom_back_right << " "; print5bitbin(get5bitCorner(bottom_back_right )); cout << "\n";
-//        cout << bottom_back_left << " "; print5bitbin(get5bitCorner(bottom_back_left )); cout << "\n";
+        // cout << top_front_right << " "; print5bitbin(get5bitCorner(top_front_right )); cout  << "\n";
+        // cout << top_front_left << " "; print5bitbin(get5bitCorner(top_front_left )); cout << "\n";
+        // cout << top_back_right << " "; print5bitbin(get5bitCorner(top_back_right )); cout << "\n";
+        // cout << top_back_left  << " "; print5bitbin(get5bitCorner(top_back_left  )); cout << "\n";
+        // cout << bottom_front_right  << " "; print5bitbin(get5bitCorner(bottom_front_right  )); cout << "\n";
+        // cout << bottom_front_left << " "; print5bitbin(get5bitCorner(bottom_front_left )); cout << "\n";
+        // cout << bottom_back_right << " "; print5bitbin(get5bitCorner(bottom_back_right )); cout << "\n";
+        // cout << bottom_back_left << " "; print5bitbin(get5bitCorner(bottom_back_left )); cout << "\n";
 
         return ret;
     }
